requirements.txt
fastapi
uvicorn
flake8
pinecone-client
numpy
pydantic
